+++ code/run_mrc.py
@@ -4,6 +4,7 @@ from typing import Dict, NoReturn, Tuple
 
 from configure import *
 from preprocess import *
+from mrc_metrics import *
 from datasets import (
     load_metric,
     Value,
@@ -18,6 +19,7 @@ from transformers import (
     DataCollatorForSeq2Seq,
     TrainingArguments,
 )
+
 
 from utils_qa import postprocess_qa_predictions, check_no_error
 from trainer_qa import QuestionAnsweringTrainer
@@ -237,47 +239,7 @@ def run_generation_mrc(
         # pad_to_multiple_of=8 if training_args.fp16 else None
     )
 
-    def postprocess_text(preds, labels):
-        preds = [pred.strip() for pred in preds]
-        labels = [label.strip() for label in labels]
-        
-        # preds = ["\n".join(tokenizer.tokenize(pred)) for pred in preds]
-        # labels = ["\n".join(tokenizer.tokenize(label)) for label in labels]
-
-        return preds, labels
-
-
-    metric = load_metric("squad")
-
-    def compute_metrics(eval_preds):
-        import numpy as np
-        preds, labels = eval_preds
-
-        if isinstance(preds, tuple):
-            preds = preds[0]
-
-        max_val_samples = 16
-        # breakpoint()
-
-        # print(preds)
-        # print(labels)
-
-        decoded_preds = tokenizer.batch_decode(preds, skip_special_tokens=True)
-        labels = np.where(labels != -100, labels, tokenizer.pad_token_id)
-        
-        decoded_labels = tokenizer.batch_decode(labels, skip_special_tokens=True)
-
-        decoded_preds, decoded_labels = postprocess_text(decoded_preds, decoded_labels)
-        
-        formatted_predictions = [{"id": ex["id"], "prediction_text": decoded_preds[i]} for i, ex in         
-                                 enumerate(datasets["validation"].select(range(max_val_samples)))]
-        references = [{"id": ex["id"], "answers": ex["answers"]} for ex in datasets["validation"].select(range(max_val_samples))]
- 
-        result = metric.compute(predictions=formatted_predictions, references=references)
-
-        return result
-    
-    
+    compute_metrics = gen_metrics(tokenizer,datasets["validation"])
     print(train_dataset)
 
     trainer = Seq2SeqTrainer(
